Create a graph for HelloWorld directly on memory. Updated According to Previous peer review contents.
