+++ CMakeLists.txt
@@ -3,6 +3,13 @@ cmake_minimum_required(VERSION 3.15)
 include(CheckSymbolExists)
 include(CheckIPOSupported)
 
+find_program(CLANG_TIDY_EXE NAMES clang-tidy)
+if(CLANG_TIDY_EXE)
+  set(CMAKE_CXX_CLANG_TIDY "${CLANG_TIDY_EXE}")
+else()
+  message(WARNING "clang-tidy not found!")
+endif()
+
 option(NINJA_BUILD_BINARY "Build ninja binary" ON)
 option(NINJA_FORCE_PSELECT "Use pselect() even on platforms that provide ppoll()" OFF)
 
@@ -84,6 +91,45 @@ else()
 endif()
 target_include_directories(libninja-re2c PRIVATE src)
 
+# --- Create shared library ---
+add_library(libninja_shared SHARED
+	src/build_log.cc
+	src/build.cc
+	src/clean.cc
+	src/clparser.cc
+	src/dyndep.cc
+	src/dyndep_parser.cc
+	src/debug_flags.cc
+	src/deps_log.cc
+	src/disk_interface.cc
+	src/edit_distance.cc
+	src/elide_middle.cc
+	src/eval_env.cc
+	src/graph.cc
+	src/graphviz.cc
+	src/json.cc
+	src/line_printer.cc
+	src/manifest_parser.cc
+	src/metrics.cc
+	src/missing_deps.cc
+	src/parser.cc
+	src/state.cc
+	src/status_printer.cc
+	src/string_piece_util.cc
+	src/util.cc
+	src/version.cc
+)
+
+# Link re2c library if required
+target_link_libraries(libninja_shared PRIVATE libninja-re2c)
+
+# Include directories for the shared library
+target_include_directories(libninja_shared PUBLIC ${CMAKE_SOURCE_DIR}/src)
+
+# Set shared library properties
+set_target_properties(libninja_shared PROPERTIES OUTPUT_NAME "ninja_shared")
+
+
 # --- Check for 'browse' mode support
 function(check_platform_supports_browse_mode RESULT)
 	# Make sure the inline.sh script works on this platform.
@@ -145,6 +191,7 @@ add_library(libninja OBJECT
 	src/metrics.cc
 	src/missing_deps.cc
 	src/parser.cc
+	src/real_command_runner.cc
 	src/state.cc
 	src/status_printer.cc
 	src/string_piece_util.cc
@@ -186,6 +233,7 @@ else()
 endif()
 
 target_compile_features(libninja PUBLIC cxx_std_11)
+target_compile_features(libninja-re2c PUBLIC cxx_std_11)
 
 #Fixes GetActiveProcessorCount on MinGW
 if(MINGW)
@@ -205,6 +253,12 @@ if(NINJA_BUILD_BINARY)
 
 	if(WIN32)
 		target_sources(ninja PRIVATE windows/ninja.manifest)
+	endif()
+
+	option(NINJA_CLANG_TIDY "Run clang-tidy on source files" OFF)
+	if(NINJA_CLANG_TIDY)
+		set_target_properties(libninja PROPERTIES CXX_CLANG_TIDY "clang-tidy;--use-color")
+		set_target_properties(ninja    PROPERTIES CXX_CLANG_TIDY "clang-tidy;--use-color")
 	endif()
 endif()
 
@@ -311,8 +365,10 @@ if(BUILD_TESTING)
     target_link_options(manifest_parser_perftest PRIVATE "-Wl,-bmaxdata:0x80000000")
   endif()
 
+
   add_test(NAME NinjaTest COMMAND ninja_test)
-endif()
+
+  endif()
 
 if(NINJA_BUILD_BINARY)
 	install(TARGETS ninja)
