+++ new-graph/hello_world.cpp
@@ -1,0 +1,10 @@
+#include <stdio.h>
+
+#ifndef N
+#define N "World"
+#endif
+
+int main() {
+  printf("Hello %s! \n", N);
+  return 0;
+}
