+++ src/manifest_parser.h
@@ -45,7 +45,7 @@ struct ManifestParser : public Parser {
     return Parse("input", input, err);
   }
 
-private:
+ private:
   /// Parse a file, given its contents as a string.
   bool Parse(const std::string& filename, const std::string& input,
              std::string* err);
